+++ packages/client/src/components/list/Card.tsx
@@ -1,4 +1,3 @@
-import clsx from 'clsx';
 import React, { useEffect, useState } from 'react';
 import { BsThreeDots } from 'react-icons/bs';
 import { useContextMenu } from 'use-context-menu';
@@ -75,24 +74,6 @@ export const Card: React.FC<React.PropsWithChildren<CardProps>> = ({
 
 export const CardTags: React.FC<React.PropsWithChildren> = ({ children }) => {
   return <div className={styles.tags}>{children}</div>;
-};
-
-interface CardTagProps {
-  icon?: React.ReactNode;
-  variant?: 'default' | 'warning' | 'error';
-}
-
-export const CardTag: React.FC<React.PropsWithChildren<CardTagProps>> = ({
-  icon,
-  children,
-  variant = 'default',
-}) => {
-  return (
-    <div className={clsx(styles.tag, styles[variant])}>
-      {icon}
-      {!!children && <span>{children}</span>}
-    </div>
-  );
 };
 
 export const CardMenu: React.FC<React.PropsWithChildren> = ({ children }) => {